Correción incompatibilidad entre negocio y Base de datos. Volver a ejecutar scripts de BBDD y de datos.

+++ TpFinal/Negocio/DetalleProductosCompraNegocio.cs
@@ -29,9 +29,9 @@ namespace Negocio
                     aux.cantidad = (int)datos.Lector["cantidad"];
                     aux.codProducto = (string)datos.Lector["codProducto"];
                     aux.descripcion = (string)datos.Lector["descripcion"];
-                    aux.monto = (int)datos.Lector["monto"];
-                    aux.montoDescuento = (int)datos.Lector["montoDesc"];
-                    aux.precioVenta = (int)datos.Lector["precioVenta"];
+                    aux.monto = (decimal)datos.Lector["monto"];
+                    aux.montoDescuento = (decimal)datos.Lector["montoDesc"];
+                    aux.precioVenta = (decimal)datos.Lector["precioVenta"];
                     aux.estado = (bool)datos.Lector["estado"];
 
                     detalleProductos.Add(aux);
